AddAndEditViewをAddViewとEditViewに分ける
NavigationBarのUIを分けるため
